refactor(verification): reorganize verification module structure
- move verification decorators to separate files
- update controller path and api tags
- add index files for better module organization
