+++ src/verification/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './send-code-docs.decorator';
+export * from './verify-code-docs.decorator';
