+++ src/verification/dto/index.ts
@@ -1,0 +1,2 @@
+export { SendCodeDto } from './send-code.dto';
+export { VerifyCodeDto } from './verify-code.dto';
